backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pymysql==1.1.0
firebase-admin==6.4.0
chromadb==0.4.22
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
python-multipart==0.0.6
sentence-transformers==2.3.1
email-validator==2.1.0
pypdf==3.17.4
Pillow==10.2.0
pytesseract==0.3.10
